move tipset validator tests to the validator module
